Description Changes
Feel free to use any of these or throw them out at your leisure.  I made an initial pass just to correct typos but then thought that a few of them sounded awkward.  Most of them are almost the same, but I fixed a few grammatical issues such as "in the mak of" being changed to "in the production of", and I think I just outright rewrote a couple.

To keep the original a bit closer, "in the make of" could also be swapped for "in the making of" which would be grammatically correct.